x230: README: avoid confusion in our 8M external flashing docs
We recently added that the whole 8M chip flashing is optional, which is
important indeed. In brough some more stuff in which may be confusion.

We don't want to mention one specific OS because we're really generic
people should try to install whatever they want.

Also, "the command defaults to..." is misleading, as there are commandline
options applied in our example. Remove that.

And lastly, as long as users make the backups we always suggest in the
examples, they can change things at any time.