common: make sure to init submodules after every git checkout
Should a newly checked out ref have a submodule that we did not have
before, it would not be fetched and checked out by our download script.
The coreboot build system would handle it. But since we currently do
submodule handling we should cover that case as well.
This also aligns all three callers to use the same arguments.

Signed-off-by: Henning Schild <henning@hennsch.de>