Add config file for vgabios build for T530
Cloned "nonfree-defconfig-5243dd96de" from X230 dir to T530, and changed the line with the model. I looked through the repo to see if anything else is needed to add vgabios and found nothing.

Is this all that's required?

I've made a build myself and am ready to test in hardware if you tell me this is good.